requirements.txt
tensorflow>=2.10
streamlit
pillow
numpy
scikit-learn
pandas
requests
fastapi
uvicorn
python-multipart




